+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Problem with the `fill` operation for the Model saving process, where situation with increments will create issues
+* Export of `APIOptions` type
 
 ## [0.13.1] - 2023-02-23